Introduce source process-granted access permits
Previously each consumer had to obtain its own access permit when
attaching a segment, and this permit was valid only for this segment. To
avoid clients having to separately keep track of access permits, we add
the `XPMEM_UNIVERSAL` flag to allow the source process to create an
access permit (as per usual via `xpmem_get()`) that can be passed to
consumers to attach the segment.