Add attachment hugepage and mkwrite fault support
Allow pages in attachments to be backed by (PMD) hugepage PTEs if they
are properly aligned and the backing page in the segment is also a
hugepage.

Also add mkwrite fault support to allow ranges in a segment to be write
protected via userfaultfd's UFFDIO_WRITEPROTECT feature and receive
notifications when a previously read-only mapping in an associated
attachment is written to.