Add library function to block page faults
Add a function that can be called to temporarily block page faults to a
specified segment. Intended to be used with userfaultfd. A source
process can block page faults, mmap() over and/or register a region of
the segment with userfaultfd, and then unblock page faults to not miss
any events.