Support `VM_FAULT_FALLBACK` in `xpmem_fault()`
Previous to Linux 6.7, `vm_operation_struct`'s `huge_fault()` hook was
never called on PTE-missing faults. Instead, the fault would fallback to
a regular PTE `fault()`. In the XPMEM fault handler, we could therefore
assume that `huge_fault()` is only called on mkwrite faults, where we
already installed a read-only PMD-PTE in the regular fault handler.

In Linux 6.7 however, `huge_fault()` is now also called on PTE-missing
faults. We therefore generalize our XPMEM fault handling so that we can
return `VM_FAULT_FALLBACK` when `huge_fault()` is called but the backing
page in the segment is not a huge page.